Проект поправлен под работу через репозитории и сделана интеграция с БД

+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -6,13 +6,13 @@ import java.util.Collection;
 
 public interface UserService {
 
-    User getUser(int id);
+    User getUser(long id);
 
     Collection<User> getUsers();
 
     User createUser(User user);
 
-    User updateUser(int id, User user);
+    User updateUser(long id, User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
